Include the Todo constructor
The constructor function of `Todo` needs to be modified for this problem solution, with the addition of the `this._onDelete = this._onDelete.bind(this);` binding.  If the constructor is omitted, the student will assume they don't need to modify it, especially if the problem statement tells them to only write code that lives in `TodoList`.  An alternative would be to tell the student they need to modify the `Todo` constructor in the problem statement, but I think this works pretty well.